refactor: move transactionData class  into tx_executor.h
