+++ src/server/journal/tx_executor.cc
@@ -4,9 +4,13 @@
 
 #include "tx_executor.h"
 
+#include <absl/strings/match.h>
+
 #include "base/logging.h"
+#include "server/journal/serializer.h"
 
 using namespace std;
+using namespace facade;
 
 namespace dfly {
 
@@ -41,6 +45,90 @@ void MultiShardExecution::CancelAllBlockingEntities() {
     tx_data.second.barrier.Cancel();
     tx_data.second.block.Cancel();
   }
+}
+
+bool TransactionData::AddEntry(journal::ParsedEntry&& entry) {
+  ++journal_rec_count;
+
+  switch (entry.opcode) {
+    case journal::Op::PING:
+      is_ping = true;
+      return true;
+    case journal::Op::EXPIRED:
+    case journal::Op::COMMAND:
+      commands.push_back(std::move(entry.cmd));
+      [[fallthrough]];
+    case journal::Op::EXEC:
+      shard_cnt = entry.shard_cnt;
+      dbid = entry.dbid;
+      txid = entry.txid;
+      return true;
+    case journal::Op::MULTI_COMMAND:
+      commands.push_back(std::move(entry.cmd));
+      dbid = entry.dbid;
+      return false;
+    default:
+      DCHECK(false) << "Unsupported opcode";
+  }
+  return false;
+}
+
+bool TransactionData::IsGlobalCmd() const {
+  if (commands.size() > 1) {
+    return false;
+  }
+
+  auto& command = commands.front();
+  if (command.cmd_args.empty()) {
+    return false;
+  }
+
+  auto& args = command.cmd_args;
+  if (absl::EqualsIgnoreCase(ToSV(args[0]), "FLUSHDB"sv) ||
+      absl::EqualsIgnoreCase(ToSV(args[0]), "FLUSHALL"sv) ||
+      (absl::EqualsIgnoreCase(ToSV(args[0]), "DFLYCLUSTER"sv) &&
+       absl::EqualsIgnoreCase(ToSV(args[1]), "FLUSHSLOTS"sv))) {
+    return true;
+  }
+
+  return false;
+}
+
+TransactionData TransactionData::FromSingle(journal::ParsedEntry&& entry) {
+  TransactionData data;
+  bool res = data.AddEntry(std::move(entry));
+  DCHECK(res);
+  return data;
+}
+
+std::optional<TransactionData> TransactionReader::NextTxData(JournalReader* reader, Context* cntx) {
+  io::Result<journal::ParsedEntry> res;
+  while (true) {
+    if (res = reader->ReadEntry(); !res) {
+      cntx->ReportError(res.error());
+      return std::nullopt;
+    }
+
+    // Check if journal command can be executed right away.
+    // Expiration checks lock on master, so it never conflicts with running multi transactions.
+    if (res->opcode == journal::Op::EXPIRED || res->opcode == journal::Op::COMMAND ||
+        res->opcode == journal::Op::PING)
+      return TransactionData::FromSingle(std::move(res.value()));
+
+    // Otherwise, continue building multi command.
+    DCHECK(res->opcode == journal::Op::MULTI_COMMAND || res->opcode == journal::Op::EXEC);
+    DCHECK(res->txid > 0);
+
+    auto txid = res->txid;
+    auto& txdata = current_[txid];
+    if (txdata.AddEntry(std::move(res.value()))) {
+      auto out = std::move(txdata);
+      current_.erase(txid);
+      return out;
+    }
+  }
+
+  return std::nullopt;
 }
 
 }  // namespace dfly
